Uses openapi-generator instead of swagger-codegen
openapi-generator is a fork of swagger-codegen but it looks like it's more
widely supported. kubernetes and a lot of other big projects use
openapi-generator instead of swagger-codegen. Plus openapi-generator supports
more languages.

The functionality is the same, just with a few differences here and there.

Here is the command used to generate this client

```
openapi-generator generate \
  -i ../front/doc/public-http-api-swagger.yaml \
  -g python \
  -p packageName=lilt,projectName=lilt-python \
  --git-user-id lilt \
  --git-repo-id lilt-python \
  -o .
```